+++ CHANGES.rst
@@ -14,6 +14,8 @@ Change Log
 
 * For the *Master* election, give a priority to nodes declared in ``forced_synchro_if`` if used.
 
+* In the process description of the Web UI Application page, add information about the node_name.
+
 * Documentation updated.